requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
langchain
langchain-openai
langchain-community
langgraph-checkpoint-mongodb
langchain-core
python-dotenv
pymupdf
